FEAT Implement proof-of-concept for file upload with `rocket-multipart-form-data`
As of now, this only works for image files.
(For some reason, the MIME file-type checking fails for e.g. audio files.)

A maximum file-size can be specified, here: 10 MiB.

As of now, the uploaded files are saved to the `./tmp` directory
(which still has to be created manually). This is a probably only temporary,
since saving to database might make more sense.

Example usage:
```
curl -X POST \
  -F 'photo=@./my_image.png' \
  http://127.0.0.1:8000/api/upload_file
```